+++ src/lib/utils/options/sources/typedoc.ts
@@ -10,17 +10,14 @@ import { SORT_STRATEGIES } from "../../sort";
 import { EntryPointStrategy } from "../../entry-point";
 import { ReflectionKind } from "../../../models/reflections/kind";
 import * as Validation from "../../validation";
-import { blockTags, inlineTags, modifierTags } from "../tsdoc-defaults";
 import { getEnumKeys } from "../../enum";
-import type {
-    BundledLanguage,
-    BundledTheme,
-} from "shiki" with { "resolution-mode": "import" };
+import type { BundledTheme } from "shiki" with { "resolution-mode": "import" };
 import {
     getSupportedLanguagesWithoutAliases,
     getSupportedThemes,
 } from "../../highlighter";
 import { setDifference } from "../../set";
+import { OptionDefaults } from "../defaults";
 
 // For convenience, added in the same order as they are documented on the website.
 export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
@@ -176,27 +173,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
                 );
             }
         },
-        defaultValue: [
-            ReflectionKind[ReflectionKind.Module],
-            ReflectionKind[ReflectionKind.Namespace],
-            ReflectionKind[ReflectionKind.Enum],
-            // Not including enum member here by default
-            ReflectionKind[ReflectionKind.Variable],
-            ReflectionKind[ReflectionKind.Function],
-            ReflectionKind[ReflectionKind.Class],
-            ReflectionKind[ReflectionKind.Interface],
-            ReflectionKind[ReflectionKind.Constructor],
-            ReflectionKind[ReflectionKind.Property],
-            ReflectionKind[ReflectionKind.Method],
-            ReflectionKind[ReflectionKind.CallSignature],
-            ReflectionKind[ReflectionKind.IndexSignature],
-            ReflectionKind[ReflectionKind.ConstructorSignature],
-            ReflectionKind[ReflectionKind.Accessor],
-            ReflectionKind[ReflectionKind.GetSignature],
-            ReflectionKind[ReflectionKind.SetSignature],
-            ReflectionKind[ReflectionKind.TypeAlias],
-            ReflectionKind[ReflectionKind.Reference],
-        ],
+        defaultValue: OptionDefaults.excludeNotDocumentedKinds,
     });
     options.addDeclaration({
         name: "excludeInternal",
@@ -331,18 +308,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         name: "highlightLanguages",
         help: (i18n) => i18n.help_highlightLanguages(),
         type: ParameterType.Array,
-        defaultValue: [
-            "bash",
-            "console",
-            "css",
-            "html",
-            "javascript",
-            "json",
-            "jsonc",
-            "json5",
-            "tsx",
-            "typescript",
-        ] satisfies BundledLanguage[],
+        defaultValue: OptionDefaults.highlightLanguages,
         validate(value, i18n) {
             const invalid = setDifference(
                 value,
@@ -441,13 +407,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         name: "excludeTags",
         help: (i18n) => i18n.help_excludeTags(),
         type: ParameterType.Array,
-        defaultValue: [
-            "@override",
-            "@virtual",
-            "@privateRemarks",
-            "@satisfies",
-            "@overload",
-        ],
+        defaultValue: OptionDefaults.excludeTags,
         validate(value, i18n) {
             if (!Validation.validate([Array, Validation.isTagString], value)) {
                 throw new Error(
@@ -716,7 +676,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         name: "blockTags",
         help: (i18n) => i18n.help_blockTags(),
         type: ParameterType.Array,
-        defaultValue: blockTags,
+        defaultValue: OptionDefaults.blockTags,
         validate(value, i18n) {
             if (!Validation.validate([Array, Validation.isTagString], value)) {
                 throw new Error(
@@ -729,7 +689,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         name: "inlineTags",
         help: (i18n) => i18n.help_inlineTags(),
         type: ParameterType.Array,
-        defaultValue: inlineTags,
+        defaultValue: OptionDefaults.inlineTags,
         validate(value, i18n) {
             if (!Validation.validate([Array, Validation.isTagString], value)) {
                 throw new Error(
@@ -742,7 +702,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         name: "modifierTags",
         help: (i18n) => i18n.help_modifierTags(),
         type: ParameterType.Array,
-        defaultValue: modifierTags,
+        defaultValue: OptionDefaults.modifierTags,
         validate(value, i18n) {
             if (!Validation.validate([Array, Validation.isTagString], value)) {
                 throw new Error(
@@ -755,7 +715,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         name: "cascadedModifierTags",
         help: (i18n) => i18n.help_modifierTags(),
         type: ParameterType.Array,
-        defaultValue: ["@alpha", "@beta", "@experimental"],
+        defaultValue: OptionDefaults.cascadedModifierTags,
         validate(value, i18n) {
             if (!Validation.validate([Array, Validation.isTagString], value)) {
                 throw new Error(
@@ -797,7 +757,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         name: "sort",
         help: (i18n) => i18n.help_sort(),
         type: ParameterType.Array,
-        defaultValue: ["kind", "instance-first", "alphabetical"],
+        defaultValue: OptionDefaults.sort,
         validate(value, i18n) {
             const invalid = new Set(value);
             for (const v of SORT_STRATEGIES) {
@@ -827,18 +787,14 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         type: ParameterType.Array,
         defaultValue: [],
         validate(value, i18n) {
-            const invalid = new Set(value);
-            const valid = getEnumKeys(ReflectionKind);
-            for (const v of valid) {
-                invalid.delete(v);
-            }
+            const invalid = setDifference(value, getEnumKeys(ReflectionKind));
 
             if (invalid.size !== 0) {
                 throw new Error(
                     i18n.option_0_specified_1_but_only_2_is_valid(
                         `kindSortOrder`,
                         Array.from(invalid).join(", "),
-                        valid.join(", "),
+                        getEnumKeys(ReflectionKind).join(", "),
                     ),
                 );
             }
@@ -929,18 +885,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
                 }
             }
         },
-        defaultValue: [
-            ReflectionKind[ReflectionKind.Enum],
-            ReflectionKind[ReflectionKind.EnumMember],
-            ReflectionKind[ReflectionKind.Variable],
-            ReflectionKind[ReflectionKind.Function],
-            ReflectionKind[ReflectionKind.Class],
-            ReflectionKind[ReflectionKind.Interface],
-            ReflectionKind[ReflectionKind.Property],
-            ReflectionKind[ReflectionKind.Method],
-            ReflectionKind[ReflectionKind.Accessor],
-            ReflectionKind[ReflectionKind.TypeAlias],
-        ],
+        defaultValue: OptionDefaults.requiredToBeDocumented,
     });
 
     options.addDeclaration({